+++ Backend/src/model/Playlist.py
@@ -1,7 +1,7 @@
 import json
 from dataclasses import dataclass
 
-import src.services.song_service as song_service
+import src.services.song_services.song_service_aws_lambda as song_service_aws_lambda
 
 
 @dataclass
@@ -15,7 +15,7 @@ class Playlist:
 
     def add_songs(self, song_names: str) -> None:
         [
-            self.songs.append(song_service.get_song(song_name))
+            self.songs.append(song_service_aws_lambda.get_song(song_name))
             for song_name in song_names
         ]
