+++ Backend/src/routers/canciones.py
@@ -3,7 +3,7 @@ from typing import Annotated, Union
 
 import src.services.dto_service as dto_service
 import src.services.security_service as security_service
-import src.services.song_service as song_service_streaming
+import src.services.song_services.song_service_aws_lambda as song_service_streaming
 from fastapi import APIRouter, Header, HTTPException, UploadFile
 from fastapi.responses import Response
 from src.model.Genre import Genre
