+++ Backend/src/services/search_service.py
@@ -2,7 +2,7 @@ import json
 
 import src.services.artist_service as artist_service
 import src.services.playlist_service as playlist_service
-import src.services.song_service as song_service
+import src.services.song_services.song_service_aws_lambda as song_service_aws_lambda
 import src.services.user_service as user_service
 from fastapi import HTTPException
 from src.services.utils import checkValidParameterString
@@ -41,7 +41,7 @@ def search_by_name(name: str) -> str:
     items = {}
 
     try:
-        songs = song_service.search_by_name(name)
+        songs = song_service_aws_lambda.search_by_name(name)
         playlists = playlist_service.search_by_name(name)
         users = user_service.search_by_name(name)
         artists = artist_service.search_by_name(name)
